Add UDS service base + refactoring of UDS/XCP classes
All DoIP message are implemented as Ethernet transport on UDS to ensure interoperability with others protocols (XCP etc...)
Routine control has been added as an example on how to implement a basic service but code should allow byteswapping in future
for fuzzing purpose.

NB: A choice should be made between usage or not of ctypes library for readability/ease of use.
NB2: Enum should be reconsidered or rewritten to handle invalid value (for dumb fuzzing)