Add SOME/IP and SOME/IP-SD protocols implementation
A quick implementation of SOME/IP is done to provide an easy way to craft SOME/IP packets.
SOME/IP is a kind of RPC mechanism done to answer requirement of tight embedded ressources.
SOME/IP-SD is a variant used for service discovery and could be used to check which services ECU is proposing.